check if dir exists before rm

+++ publish/prepareGithubPages.py
@@ -1,12 +1,10 @@
-# TODO create page/
-# TODO copy 00 - 30 to page/
 # TODO v2: replace callout syntax to work with adminition plugin
 
 import shutil
 from pathlib import Path
 
-# TODO remove old pages dir first
-shutil.rmtree('publish/pages')
+if Path('publish/pages').is_dir():
+    shutil.rmtree('publish/pages')
 # Path('publish/pages').mkdir()
 dirsToCopy = ['00 Meta', '10 Example Data',
               '20 Dataview Queries', '30 Dataview Resources']
